- Trying to replace the std with jkr. - Improvements adn additions at sring. - Changes in the bytecode for small code. - Extending register count to 16. - Working on the standard library.

+++ jkr/CodeFile/Object.h
@@ -19,7 +19,7 @@ struct FieldHeader {
 };
 
 struct ObjectHeader {
-    unsigned short CountOfFields;
+    Byte CountOfFields;
     // FieldHeader Fields[CountOfFields];
 };
